Babel + modules import/export
