+++ app/index.html
@@ -57,7 +57,7 @@
   </div>
 </footer>
 
-<script src="js/scripts.js"></script>
+<script src="app.bundle.js"></script>
 
 </body>
 </html>
