+++ app/js/api.js
@@ -1,7 +1,7 @@
 const apiKey = 'eabd967104da4e07a9c41b1342a889b1';
 const host = 'https://newsapi.org/v2';
 
-class ResponseApi {
+export default class ResponseApi{
   static getSourcesOnCategory(category){
     const url = `${host}/sources?category=${category}&apiKey=${apiKey}`;
     return fetch(url).then(response => response.json());
