+++ app/js/news.js
@@ -1,5 +1,12 @@
+function renderNews(elemId, news) {
+  let elem = elemId;
+  if (elem && news.status == 'ok') {
+    elem.innerHTML = news.articles.reduce((txt, li) => txt + `<li class='col-xs-12 col-sm-12 col-lg-6 col-xl-6'>${render(li)}</li>`, `<ul class="articles row">`) + `</ul>`;
+  }
+}
+
 function render(article) {
-    return `<div class="card">
+  return `<div class="card">
           <img class="card-img-top w-100" src="${article.urlToImage}" alt="${article.title}">
           <div class="card-block">
             <h4 class="card-title">${article.title}</h4>
@@ -10,3 +17,4 @@ function render(article) {
         </div>`;
 }
 
+export{renderNews}
