--- app/js/selectsources.js
@@ -1,19 +1,0 @@
-const newCategories = ['business', 'entertainment', 'general', 'health', 'science', 'sports', 'technology'];
-const sourcesContainer = document.querySelector('#news-sources');
-
-let htmlSelect = `
-  <select class="form-control sel-cat" id="cat_val" name="cat-select">
-    ${newCategories.reduce((allOpt, opt) => allOpt + `<option value="${opt}">${opt}</option>`, `<option value="0">Select news category:</option>`)}
-  </select>`;
-
-document.getElementById('news-categories').innerHTML = htmlSelect;
-
-document.querySelector('.sel-cat').addEventListener('change', e => {
-  let selectedCategory = e.target.value;
-
-  ResponseApi.getSourcesOnCategory(selectedCategory)
-    .then(sources => {
-      renderSources(sourcesContainer, sources);
-    });
-});
-
